Refactor Whisper decode and driver components:
- Rename `Decoded` to `DecodedTokens` across all relevant modules.
- Migrate `SpeechRegion` and utilities to a dedicated `speech_region` module; update related imports.
- Replace inline logic for padding and region validation with reusable module-level functions.
- Simplify `voice_activity_filter` by removing redundant definitions of `SpeechRegion` and its utilities.
