+++ crates/public/bunsen/CHANGELOG.md
@@ -41,7 +41,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   `WhisperFallbackConfig` (the temperature ladder, the compression-ratio / log-probability / no-speech thresholds,
   `best_of`), its clauses as pure functions (`needs_fallback`, `should_skip`, `resets_prompt`),
   `compression_ratio` (zlib, as upstream), and `decode_with_fallback`, the ladder as an orchestration over a decode
-  closure. `Decoded` is what a decode now says about an audio beyond its ids: the cumulative log probability
+  closure. `DecodedTokens` is what a decode now says about an audio beyond its ids: the cumulative log probability
   (`avg_logprob`) and the `<|nospeech|>` probability probed at the sot position of the first forward, through
   `decode_windows_full` /
   `decode_features_full`; `DecodeConfig` takes `temperature` and
